+++ app/src/main/res/layout/activity_welcome.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Welcome">
+    tools:context=".controller.Welcome">
 
     <Button
         android:id="@+id/login_button"
